backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyPDF2==3.0.1
reportlab==4.0.8
diff-match-patch==20230430
openai>=1.12.0
# python-dotenv==1.0.9
python-dotenv==1.1.0
pytest==7.4.3
# MongoDB dependencies
pymongo==4.6.2
motor==3.3.2
beanie==1.25.0
# Use pre-built packages to avoid Rust compilation
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.8
alembic==1.12.0
# Add email validation without requiring pydantic_core
email-validator==2.0.0 